advanced c functions

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	int num, prm;
+	long int num, prm;
 
 	num = 612852475143;
